+++ generator/lib/src/code_chunks.dart
@@ -76,7 +76,6 @@ class CodeChunks {
     final obxFlutter = pubspec?.hasObxFlutterDependency ?? false;
     return '''${obxFlutter ? 'Future<$obx.Store>' : '$obx.Store'} openStore(
         {String? directory,
-          String? inMemoryIdentifier,
           int? maxDBSizeInKB,
           int? maxDataSizeInKB,
           int? fileMode,
@@ -86,7 +85,6 @@ class CodeChunks {
         ${obxFlutter ? 'await loadObjectBoxLibraryAndroidCompat();' : ''}
         return $obx.Store(getObjectBoxModel(),
             directory: directory${obxFlutter ? ' ?? (await defaultStoreDirectory()).path' : ''},
-            inMemoryIdentifier: inMemoryIdentifier,
             maxDBSizeInKB: maxDBSizeInKB,
             maxDataSizeInKB: maxDataSizeInKB,
             fileMode: fileMode,